Rename createStubRequest
