+++ src/create-stub-request.js
@@ -8,14 +8,14 @@ import { isObject, isFunction, identity } from 'lodash'
  * Unlike {@link createRequest}, these action creators do not make real API calls but rather
  * resolve immediately with the provided data.
  *
- * @name stubRequest
+ * @name createStubRequest
  * @param {String} type - A unique key that will be used to identify the request internally in redux
  * @param {Object|Function} dataDefinition - Data that the request will resolve with, or a function that returns data to resolve with.
  * @returns {Function} An action creator that passes its arguments to `dataDefinition` and makes the resulting stubbed API request.
  * @example
  *
  *
- * export const fetchUser = stubRequest('FETCH_USER', (id) => ({ id }))
+ * export const fetchUser = createStubRequest('FETCH_USER', (id) => ({ id }))
  *
  * fetchUsers(5)
  * // -> won't make any api request, but will resolve with data { id: 5 }
@@ -34,7 +34,7 @@ function createActionOptions (definition, args) {
     : definition
 }
 
-function stubRequest (type, definition=identity) {
+function createStubRequest (type, definition=identity) {
   if (!type) throw new Error('Must include a type for your request.')
   if (!(isObject(definition) || isFunction(definition))) throw new Error('Request definition must be an object or a function.')
   function actionCreator (...args) {
@@ -50,4 +50,4 @@ function stubRequest (type, definition=identity) {
   return actionCreator
 }
 
-export default stubRequest
+export default createStubRequest
