+++ test/middleware.test.js
@@ -158,6 +158,6 @@ test('middleware dispatches default failure action with correct data argument', 
   return store.dispatch(actionWithURL(failureUrl)).catch(() => {
     const dispatchedActions = store.getActions()
     // Internal FAILURE action
-    expect(dispatchedActions[3].payload.data).toEqual({ status: 401, data: { url: failureUrl }})
+    expect(dispatchedActions[3].payload.data).toEqual({ status: 422, data: { url: failureUrl }})
   })
 })
